server/requirements.txt
amqp==5.2.0
anyio==4.7.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
channels==4.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
distlib==0.3.9
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
h11==0.14.0
httptools==0.6.4
idna==3.10
kombu==5.4.2
pbr==6.1.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
stevedore==5.3.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
vine==5.1.0
virtualenv==20.26.6
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
watchfiles==1.0.3
wcwidth==0.2.13
websockets==14.1